create library file for kernel, move all definitions to the new library file, tested all functionality and regression tests successfully
